# Make TensorIterator stop promoting types by copying