--- a/aten/src/ATen/native/cpu/BinaryOpsKernel.cpp
+++ b/aten/src/ATen/native/cpu/BinaryOpsKernel.cpp
@@ -14,13 +14,13 @@ namespace {
 using namespace vec256;
 
 void add_kernel(TensorIterator& iter, Scalar alpha_scalar) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
       using scalar_t = bool;
       auto alpha = alpha_scalar.to<scalar_t>();
       cpu_kernel(iter,
         [=](scalar_t a, scalar_t b) -> scalar_t { return a + alpha * b; });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND(kBFloat16, iter.dtype(), "add_cpu/sub_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND(kBFloat16, iter.common_dtype(), "add_cpu/sub_cpu", [&]() {
       auto alpha = alpha_scalar.to<scalar_t>();
       auto alpha_vec = Vec256<scalar_t>(alpha);
       cpu_kernel_vec(iter,
@@ -30,10 +30,10 @@ void add_kernel(TensorIterator& iter, Scalar alpha_scalar) {
         });
       });
   }
-} 
+}
 
 void atan2_kernel(TensorIterator& iter) {
-  AT_DISPATCH_FLOATING_TYPES(iter.dtype(), "atan2_cpu", [&]() {
+  AT_DISPATCH_FLOATING_TYPES(iter.common_dtype(), "atan2_cpu", [&]() {
     cpu_kernel_vec(iter, [=](scalar_t a, scalar_t b) -> scalar_t {
     return std::atan2(a, b);
   },
@@ -48,10 +48,10 @@ void sub_kernel(TensorIterator& iter, Scalar alpha_scalar) {
 }
 
 void mul_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     cpu_kernel(iter, [=](bool a, bool b) -> bool { return a && b; });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND(kBFloat16, iter.dtype(), "mul_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND(kBFloat16, iter.common_dtype(), "mul_cpu", [&]() {
       cpu_kernel_vec(iter,
         [=](scalar_t a, scalar_t b) -> scalar_t { return a * b; },
         [=](Vec256<scalar_t> a, Vec256<scalar_t> b) {
@@ -62,16 +62,16 @@ void mul_kernel(TensorIterator& iter) {
 }
 
 void div_kernel(TensorIterator& iter) {
-  if (isIntegralType(iter.dtype(), /*includeBool*/ false)) {
+  if (isIntegralType(iter.common_dtype(), /*includeBool*/ false)) {
     // There's no SIMD integer division, so don't try to vectorize it.
     // TODO: if the divisor is a scalar, rewrite as multiplication by a constant.
-    AT_DISPATCH_INTEGRAL_TYPES(iter.dtype(), "div_cpu", [&]() {
+    AT_DISPATCH_INTEGRAL_TYPES(iter.common_dtype(), "div_cpu", [&]() {
       cpu_kernel(iter, [](scalar_t a, scalar_t b) -> scalar_t {
         return a / b;
       });
     });
-  } else if (isComplexType(iter.dtype())) {
-      AT_DISPATCH_COMPLEX_TYPES(iter.dtype(), "div_cpu", [&]() {
+  } else if (isComplexType(iter.common_dtype())) {
+      AT_DISPATCH_COMPLEX_TYPES(iter.common_dtype(), "div_cpu", [&]() {
         cpu_kernel_vec(iter,
           [=](scalar_t a, scalar_t b) __ubsan_ignore_float_divide_by_zero__ -> scalar_t {
              return a / b;
@@ -81,7 +81,7 @@ void div_kernel(TensorIterator& iter) {
           });
       });
     } else {
-    AT_DISPATCH_FLOATING_TYPES_AND(kBFloat16, iter.dtype(), "div_cpu", [&]() {
+    AT_DISPATCH_FLOATING_TYPES_AND(kBFloat16, iter.common_dtype(), "div_cpu", [&]() {
       cpu_kernel_vec(iter,
         [=](scalar_t a, scalar_t b) __ubsan_ignore_float_divide_by_zero__ -> scalar_t {
            return a / b;
@@ -94,7 +94,7 @@ void div_kernel(TensorIterator& iter) {
 }
 
 void logical_xor_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND3(kBool, kBFloat16, kHalf, iter.input_dtype(), "logical_xor_cpu", [&]() {
       cpu_kernel(iter,
         [](scalar_t a, scalar_t b) -> bool {
@@ -102,7 +102,7 @@ void logical_xor_kernel(TensorIterator& iter) {
         });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND2(kBFloat16, kHalf, iter.dtype(), "logical_xor_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND2(kBFloat16, kHalf, iter.common_dtype(), "logical_xor_cpu", [&]() {
       cpu_kernel(iter,
         [](scalar_t a, scalar_t b) -> scalar_t {
           return static_cast<scalar_t>(bool(a) != bool(b));
@@ -112,7 +112,7 @@ void logical_xor_kernel(TensorIterator& iter) {
 }
 
 void lt_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "lt_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -120,7 +120,7 @@ void lt_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "lt_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "lt_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a < b;
@@ -130,7 +130,7 @@ void lt_kernel(TensorIterator& iter) {
 }
 
 void le_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "le_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -138,7 +138,7 @@ void le_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "le_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "le_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a <= b;
@@ -148,7 +148,7 @@ void le_kernel(TensorIterator& iter) {
 }
 
 void gt_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "gt_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -156,7 +156,7 @@ void gt_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "gt_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "gt_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a > b;
@@ -166,7 +166,7 @@ void gt_kernel(TensorIterator& iter) {
 }
 
 void ge_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "ge_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -174,7 +174,7 @@ void ge_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "ge_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "ge_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a >= b;
@@ -184,7 +184,7 @@ void ge_kernel(TensorIterator& iter) {
 }
 
 void eq_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "eq_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -192,7 +192,7 @@ void eq_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "eq_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "eq_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a == b;
@@ -202,7 +202,7 @@ void eq_kernel(TensorIterator& iter) {
 }
 
 void ne_kernel(TensorIterator& iter) {
-  if (iter.dtype() == ScalarType::Bool) {
+  if (iter.common_dtype() == ScalarType::Bool) {
     AT_DISPATCH_ALL_TYPES_AND2(kBool, kBFloat16, iter.input_dtype(), "ne_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> bool {
@@ -210,7 +210,7 @@ void ne_kernel(TensorIterator& iter) {
        });
     });
   } else {
-    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.dtype(), "ne_cpu", [&]() {
+    AT_DISPATCH_ALL_TYPES_AND(kBFloat16, iter.common_dtype(), "ne_cpu", [&]() {
       cpu_kernel(iter,
        [=](scalar_t a, scalar_t b) -> scalar_t {
          return a != b;