--- a/aten/src/ATen/test/tensor_iterator_test.cpp
+++ b/aten/src/ATen/test/tensor_iterator_test.cpp
@@ -188,8 +188,9 @@ TEST(TensorIteratorTest, ComputeCommonDTypeInputOnly) {
   iter.compute_common_dtype_only_for_inputs();
   iter.build();
   EXPECT_TRUE(iter.dtype(0) == at::kBool);
-  EXPECT_TRUE(iter.dtype(1) == at::kDouble);
+  EXPECT_TRUE(iter.dtype(1) == at::kFloat);
   EXPECT_TRUE(iter.dtype(2) == at::kDouble);
+  EXPECT_TRUE(iter.common_dtype() == at::kDouble);
 }
 
 TEST(TensorIteratorTest, DoNotComputeCommonDTypeInputOnly) {