@VitalyFedyunin 2019-10-21
Please also include CPU benchmarks with and without type promotions