@zasdfgbnm 2019-10-21
@VitalyFedyunin There is very little change in the performance.

The benchmark is as follows:
```python
import torch
print(torch.__version__)
print(torch.version.git_version)

_100M = 100 * 1024 ** 2
r = torch.randn(_100M, dtype=torch.float32, device='cpu')
d = torch.randn(_100M, dtype=torch.float64, device='cpu')
%timeit r.add_(d);
```
before
```
1.4.0a0+f6c0a89
f6c0a89acc929427e2e02a8215e7692455271178
129 ms ± 5.52 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
after
```
1.4.0a0+1c1c778
1c1c778b69e7473f2ddb6dbf69c8e626f33831ff
121 ms ± 1.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```