@zasdfgbnm 2019-10-21
Without promotion on CPU:
```python
import torch
print(torch.__version__)
print(torch.version.git_version)

_100M = 100 * 1024 ** 2
a = torch.randn(_100M, dtype=torch.float32, device='cpu')
b = torch.randn(_100M, dtype=torch.float32, device='cpu')
%timeit a.add_(b);
```
before
```
1.4.0a0+f6c0a89
f6c0a89acc929427e2e02a8215e7692455271178
23.3 ms ± 158 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
after
```
1.4.0a0+1c1c778
1c1c778b69e7473f2ddb6dbf69c8e626f33831ff
23.1 ms ± 97.4 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```