@zasdfgbnm 2019-10-21
The case when there is no promotion is dispatched to the original code at https://github.com/pytorch/pytorch/pull/28344/files#diff-0d1178f1a4ce15aeb760d251974e6924R242
and
https://github.com/pytorch/pytorch/pull/28344/files#diff-baaaf6b9adceeef7f820bc73f6e49a44R161
So there shouldn't be any performance change.